remove extra actions from PdfSidebar and move handleCategoryToggle to SideBarCategories
